pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "novadb"
version = "0.1.0"
description = "AlphaFold3-style dataset curation and feature engineering pipeline for structure prediction training"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "NovaDB Team", email = "contact@novadb.org"}
]
maintainers = [
    {name = "NovaDB Team", email = "contact@novadb.org"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
keywords = [
    "alphafold",
    "alphafold3",
    "protein structure",
    "structure prediction",
    "machine learning",
    "deep learning",
    "bioinformatics",
    "dataset curation",
    "feature engineering",
    "molecular biology",
]

dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    
    # Bioinformatics
    "biopython>=1.81",
    "rdkit>=2023.3.1",
    
    # Data handling
    "h5py>=3.8.0",
    "pyarrow>=12.0.0",
    "msgpack>=1.0.0",
    
    # Cloud storage
    "boto3>=1.28.0",  # AWS S3
    "google-cloud-storage>=2.10.0",  # Google Cloud Storage
    "azure-storage-blob>=12.17.0",  # Azure Blob Storage
    
    # Configuration and CLI
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    
    # Parallel processing
    "joblib>=1.3.0",
    "ray>=2.5.0",
    
    # Logging and monitoring
    "structlog>=23.1.0",
    
    # HTTP and networking
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    "httpx>=0.24.0",
    
    # Async S3
    "aioboto3>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
    "pre-commit>=3.3.0",
]

performance = [
    # Optional performance optimizations
    "numba>=0.58.0",
]

distributed = [
    # Distributed processing with Redis
    "redis[hiredis]>=5.0.0",
]

hmmer = [
    # HMMER tools are installed separately via conda or system package manager
    # This is a placeholder for documentation
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=2.0.0",
]

all = [
    "novadb[dev,docs,performance,distributed]",
]

[project.scripts]
novadb = "novadb.cli:main"
novadb-pipeline = "novadb.pipeline.runner:main"
novadb-download = "novadb.data.download:main"

[project.urls]
Homepage = "https://github.com/novadb/novadb"
Documentation = "https://novadb.readthedocs.io"
Repository = "https://github.com/novadb/novadb.git"
Issues = "https://github.com/novadb/novadb/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["D100", "D104"]
target-version = "py310"

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_external: marks tests that require external tools (HMMER, etc.)",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/novadb"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
fail_under = 70

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["novadb"]
skip_gitignore = true
extend_skip = [".venv", "venv", "build", "dist"]

[tool.bandit]
exclude_dirs = ["tests", "docs"]
skips = ["B101"]  # Skip assert_used check (used in tests)

[tool.bandit.assert_used]
skips = ["*_test.py", "*test_*.py"]
